refactor(cli): hybrid src layout and version guard test
Group actions under actions/{reporting,graph,explore,health}, shared
helpers under internal/, keep cli entry at src root. Add
version.test.ts so CLI_PACKAGE_VERSION stays aligned with package.json.

Made-with: Cursor
