+++ packages/dbt-tools/cli/src/actions/explore/inventory-action.test.ts
@@ -5,7 +5,7 @@ import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import {
   createJaffleArtifactBundleDir,
   createJaffleManifestOnlyDir,
-} from "./cli-test-bundle-dir";
+} from "../../internal/cli-test-bundle-dir";
 import { inventoryAction, formatInventory } from "./inventory-action";
 
 describe("inventoryAction", () => {
