+++ packages/dbt-tools/cli/src/actions/explore/search-action.ts
@@ -16,8 +16,8 @@ import {
 import {
   resolveCliArtifactPaths,
   type ArtifactRootCliOptions,
-} from "./cli-artifact-resolve";
-import { applyListPaging } from "./cli-pagination";
+} from "../../internal/cli-artifact-resolve";
+import { applyListPaging } from "../../internal/cli-pagination";
 
 export type SearchOptions = {
   type?: string;
