+++ packages/dbt-tools/cli/src/actions/graph/export-action.ts
@@ -17,7 +17,7 @@ import {
 import {
   resolveCliArtifactPaths,
   type ArtifactRootCliOptions,
-} from "./cli-artifact-resolve";
+} from "../../internal/cli-artifact-resolve";
 
 export type ExportCliOptions = {
   format?: string;
