+++ packages/dbt-tools/cli/src/actions/graph/impact-action.ts
@@ -19,7 +19,7 @@ import {
 import {
   resolveCliArtifactPaths,
   type ArtifactRootCliOptions,
-} from "./cli-artifact-resolve";
+} from "../../internal/cli-artifact-resolve";
 
 export type ImpactCliOptions = {
   fields?: string;
