+++ packages/dbt-tools/cli/src/actions/health/diagnose-action.ts
@@ -14,7 +14,7 @@ import {
 import {
   resolveCliArtifactPaths,
   type ArtifactRootCliOptions,
-} from "./cli-artifact-resolve";
+} from "../../internal/cli-artifact-resolve";
 
 export type DiagnoseCliOptions = {
   fields?: string;
