+++ packages/dbt-tools/cli/src/actions/health/status-action.ts
@@ -18,7 +18,7 @@ import {
   resolveCliArtifactPaths,
   resolveEffectiveDbtTarget,
   type ArtifactRootCliOptions,
-} from "./cli-artifact-resolve";
+} from "../../internal/cli-artifact-resolve";
 
 export type StatusOptions = {
   json?: boolean;
