+++ packages/dbt-tools/cli/src/actions/reporting/failures-action.test.ts
@@ -6,7 +6,7 @@ import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import {
   createJaffleArtifactBundleDir,
   createJaffleRunResultsOnlyDir,
-} from "./cli-test-bundle-dir";
+} from "../../internal/cli-test-bundle-dir";
 import { failuresAction } from "./failures-action";
 
 describe("failuresAction", () => {
