+++ packages/dbt-tools/cli/src/actions/reporting/timeline-action.ts
@@ -22,7 +22,7 @@ import {
 import {
   resolveCliArtifactPaths,
   type ArtifactRootCliOptions,
-} from "./cli-artifact-resolve";
+} from "../../internal/cli-artifact-resolve";
 
 export type TimelineOptions = {
   sort?: string;
