+++ packages/dbt-tools/cli/src/cli-actions.ts
@@ -1,12 +1,15 @@
-export { runReportAction } from "./run-report-action";
-export { depsAction } from "./deps-action";
-export { inventoryAction } from "./inventory-action";
-export { timelineAction } from "./timeline-action";
-export { failuresAction } from "./failures-action";
-export { searchAction } from "./search-action";
-export { discoverAction } from "./discover-action";
-export { explainAction } from "./explain-action";
-export { impactAction } from "./impact-action";
-export { diagnoseRunAction, diagnoseNodeAction } from "./diagnose-action";
-export { exportAction } from "./export-action";
-export { statusAction } from "./status-action";
+export { runReportAction } from "./actions/reporting/run-report-action";
+export { depsAction } from "./actions/graph/deps-action";
+export { inventoryAction } from "./actions/explore/inventory-action";
+export { timelineAction } from "./actions/reporting/timeline-action";
+export { failuresAction } from "./actions/reporting/failures-action";
+export { searchAction } from "./actions/explore/search-action";
+export { discoverAction } from "./actions/explore/discover-action";
+export { explainAction } from "./actions/graph/explain-action";
+export { impactAction } from "./actions/graph/impact-action";
+export {
+  diagnoseRunAction,
+  diagnoseNodeAction,
+} from "./actions/health/diagnose-action";
+export { exportAction } from "./actions/graph/export-action";
+export { statusAction } from "./actions/health/status-action";
