+++ packages/dbt-tools/cli/src/internal/cli-pagination.ts
